include eval and metrics for Tapex
