+++ primeqa/tableqa/tapex/models/tapex_model.py
@@ -33,6 +33,8 @@ from transformers.trainer_utils import get_last_checkpoint, is_main_process
 from transformers.utils import check_min_version
 from primeqa.tableqa.tapex.preprocessors.wikisql import preprocess_tableqa_function_wikisql
 from primeqa.tableqa.tapex.preprocessors.wikitablequestions import preprocess_tableqa_function_wtq
+from primeqa.tableqa.tapex.run_tapex import ModelArguments, DataTrainingArguments
+from primeqa.tableqa.tapex.metrics.tapex_accuracy import TapexAccuracy
 
 from primeqa.tableqa.tapex.run_tapex import ModelArguments, DataTrainingArguments, compute_metrics
 logger = logging.getLogger(__name__)
@@ -151,31 +153,6 @@ class TapexModel():
 
         # Set seed before initializing model.
         set_seed(training_args.seed)                                                      
-    
-        # if data_args.dataset_name is not None:
-        #     # Downloading and loading a dataset from the hub.
-        #     datasets = load_dataset(data_args.dataset_name, data_args.dataset_config_name, cache_dir=model_args.cache_dir)
-        # else:
-        #     data_files = {}
-        #     if data_args.train_file is not None:
-        #         data_files["train"] = data_args.train_file
-        #         extension = data_args.train_file.split(".")[-1]
-        #     if data_args.validation_file is not None:
-        #         data_files["validation"] = data_args.validation_file
-        #         extension = data_args.validation_file.split(".")[-1]
-        #     if data_args.test_file is not None:
-        #         data_files["test"] = data_args.test_file
-        #         extension = data_args.test_file.split(".")[-1]
-        #     datasets = load_dataset(extension, data_files=data_files, cache_dir=model_args.cache_dir)
-
-        # See more about loading any type of standard or custom dataset (from files, python dict, pandas DataFrame, etc) at
-        # https://huggingface.co/docs/datasets/loading_datasets.html.
-
-        # Load pretrained model and tokenizer
-        #
-        # Distributed training:
-        # The .from_pretrained methods guarantee that only one local process can concurrently
-        # download model & vocab.
 
         config = AutoConfig.from_pretrained(
             model_args.config_name if model_args.config_name else model_args.model_name_or_path,
@@ -200,10 +177,6 @@ class TapexModel():
             add_prefix_space=True,
         )
 
-        # if data_args.dataset_name == 'wikisql':
-        #     preprocessor = WikiSQLProcessor(tokenizer,data_args)
-        # elif data_args.dataset_name == 'wikitablequestions':
-        #     preprocessor = WTQProcessor(tokenizer,data_args)
 
         # load Bart based Tapex model (default tapex-large)
         self._model = BartForConditionalGeneration.from_pretrained(
@@ -249,12 +222,6 @@ class TapexModel():
             if data_args.train_file is not None:
                 data_files["train"] = data_args.train_file
                 extension = data_args.train_file.split(".")[-1]
-            if data_args.validation_file is not None:
-                data_files["validation"] = data_args.validation_file
-                extension = data_args.validation_file.split(".")[-1]
-            if data_args.test_file is not None:
-                data_files["test"] = data_args.test_file
-                extension = data_args.test_file.split(".")[-1]
             datasets = load_dataset(extension, data_files=data_files, cache_dir=model_args.cache_dir)
         
         column_names = datasets["train"].column_names
@@ -297,60 +264,8 @@ class TapexModel():
             pad_to_multiple_of=8 if training_args.fp16 else None,
         )
 
-        def postprocess_text(preds, labels):
-            preds = [pred.strip() for pred in preds]
-            labels = [label.strip() for label in labels]
-
-            return preds, labels
-
-        def compute_metrics(eval_preds):
-            preds, labels = eval_preds
-            if isinstance(preds, tuple):
-                preds = preds[0]
-            decoded_preds = self._tokenizer.batch_decode(preds, skip_special_tokens=True)
-            if data_args.ignore_pad_token_for_loss:
-                # Replace -100 in the labels as we can't decode them.
-                labels = np.where(labels != -100, labels, tokenizer.pad_token_id)
-            decoded_labels = self._tokenizer.batch_decode(labels, skip_special_tokens=True)
-
-            # Some simple post-processing
-            decoded_preds, decoded_labels = postprocess_text(decoded_preds, decoded_labels)
-
-            delimiter = ", "
-
-            # define example evaluation
-            def evaluate_example(predict_str: str, ground_str: str):
-                predict_spans = predict_str.split(delimiter)
-                ground_spans = ground_str.split(delimiter)
-                predict_values = defaultdict(lambda: 0)
-                ground_values = defaultdict(lambda: 0)
-                for span in predict_spans:
-                    try:
-                        predict_values[float(span)] += 1
-                    except ValueError:
-                        predict_values[span.strip()] += 1
-                for span in ground_spans:
-                    try:
-                        ground_values[float(span)] += 1
-                    except ValueError:
-                        ground_values[span.strip()] += 1
-                _is_correct = predict_values == ground_values
-                return _is_correct
-
-            def get_denotation_accuracy(predictions: List[str], references: List[str]):
-                assert len(predictions) == len(references)
-                correct_num = 0
-                for predict_str, ground_str in zip(predictions, references):
-                    is_correct = evaluate_example(predict_str.lower(), ground_str.lower())
-                    if is_correct:
-                        correct_num += 1
-                return correct_num / len(predictions)
-
-            accuracy = get_denotation_accuracy(decoded_preds, decoded_labels)
-            result = {"denotation_accuracy": accuracy}
-
-            return result
-
+        training_args.predict_with_generate=True
+        tf = TapexAccuracy(self._tokenizer,data_args)
         # Initialize our Trainer
         trainer = Seq2SeqTrainer(
             model=self._model,
@@ -358,7 +273,7 @@ class TapexModel():
             train_dataset=train_dataset,
             tokenizer=self._tokenizer,
             data_collator=data_collator,
-            compute_metrics=compute_metrics if training_args.predict_with_generate else None,
+            compute_metrics=tf.compute_metrics if training_args.predict_with_generate else None,
         )
 
         checkpoint = None
@@ -374,6 +289,8 @@ class TapexModel():
             data_args.max_train_samples if data_args.max_train_samples is not None else len(train_dataset)
         )
         metrics["train_samples"] = min(max_train_samples, len(train_dataset))
+        
+        print("train_samples ", metrics["train_samples"])
 
         trainer.log_metrics("train", metrics)
         trainer.save_metrics("train", metrics)
@@ -437,88 +354,28 @@ class TapexModel():
             pad_to_multiple_of=8 if training_args.fp16 else None,
         )
 
-        def postprocess_text(preds, labels):
-            preds = [pred.strip() for pred in preds]
-            labels = [label.strip() for label in labels]
-
-            return preds, labels
-
-        def compute_metrics(eval_preds):
-            preds, labels = eval_preds
-            if isinstance(preds, tuple):
-                preds = preds[0]
-            decoded_preds = self._tokenizer.batch_decode(preds, skip_special_tokens=True)
-            if data_args.ignore_pad_token_for_loss:
-                # Replace -100 in the labels as we can't decode them.
-                labels = np.where(labels != -100, labels, tokenizer.pad_token_id)
-            decoded_labels = self._tokenizer.batch_decode(labels, skip_special_tokens=True)
-
-            # Some simple post-processing
-            decoded_preds, decoded_labels = postprocess_text(decoded_preds, decoded_labels)
-
-            delimiter = ", "
-
-            # define example evaluation
-            def evaluate_example(predict_str: str, ground_str: str):
-                predict_spans = predict_str.split(delimiter)
-                ground_spans = ground_str.split(delimiter)
-                predict_values = defaultdict(lambda: 0)
-                ground_values = defaultdict(lambda: 0)
-                for span in predict_spans:
-                    try:
-                        predict_values[float(span)] += 1
-                    except ValueError:
-                        predict_values[span.strip()] += 1
-                for span in ground_spans:
-                    try:
-                        ground_values[float(span)] += 1
-                    except ValueError:
-                        ground_values[span.strip()] += 1
-                _is_correct = predict_values == ground_values
-                return _is_correct
-
-            def get_denotation_accuracy(predictions: List[str], references: List[str]):
-                assert len(predictions) == len(references)
-                correct_num = 0
-                for predict_str, ground_str in zip(predictions, references):
-                    is_correct = evaluate_example(predict_str.lower(), ground_str.lower())
-                    if is_correct:
-                        correct_num += 1
-                return correct_num / len(predictions)
-
-            accuracy = get_denotation_accuracy(decoded_preds, decoded_labels)
-            result = {"denotation_accuracy": accuracy}
-
-            return result
-
+        tf = TapexAccuracy(self._tokenizer,data_args)
+        training_args.predict_with_generate=True
         # Initialize our Trainer
         trainer = Seq2SeqTrainer(
             model=self._model,
             args=training_args,
-            eval_dataset=test_dataset
+            eval_dataset=test_dataset,
             tokenizer=self._tokenizer,
             data_collator=data_collator,
-            compute_metrics=compute_metrics,
+            compute_metrics=tf.compute_metrics if training_args.predict_with_generate else None,
         )
 
-        checkpoint = None
-        if training_args.resume_from_checkpoint is not None:
-            checkpoint = training_args.resume_from_checkpoint
-        elif last_checkpoint is not None:
-            checkpoint = last_checkpoint
-        train_result = trainer.train(resume_from_checkpoint=checkpoint)
-        trainer.save_model()  # Saves the tokenizer too for easy upload
+        logger.info("*** Evaluate ***")
+        print("max_eval_samples is set as: ", data_args.max_eval_samples)
 
-        metrics = train_result.metrics
-        max_train_samples = (
-            data_args.max_train_samples if data_args.max_train_samples is not None else len(train_dataset)
+        metrics = trainer.evaluate(
+            max_length=data_args.val_max_target_length, num_beams=data_args.num_beams, metric_key_prefix="eval"
         )
-        metrics["train_samples"] = min(max_train_samples, len(train_dataset))
+        max_eval_samples = data_args.max_eval_samples if data_args.max_eval_samples is not None else len(test_dataset)
+       
+        metrics["eval_samples"] = min(max_eval_samples, len(test_dataset))
 
         trainer.log_metrics("eval", metrics)
         trainer.save_metrics("eval", metrics)
-        trainer.save_state()
-
-
-
